Don't treat 30x responses as errors
It's not possible for `http.Client` to ignore a redirect response without
returning an error. Which results in the following output (nevermind the
incorrect number):

2013/10/18 13:47:56 Request error:Get https://[REDACTED]/[REDACTED]: stopped after 2 redirect

By using the lower-level `http.Transport.Roundtrip` the redirect is
completely ignored. It also won't do clever stuff with authentication and
cookies, which is desirable, because they should already be in the
constructed request.