Rolling timeout while reading HTTP bodies
The --output-http-timeout option mieaning is changed. Previously it was
a hard limit for which the whole body should have been read. But most of
the time this does not work for large bodies.

Tho timeout is now set before every Read operation effectively meaning
the body will continue to be read as long as there is any data flowing.
The timeout will be triggered only after an period of inactivity is reached.